devops-project-2974: se añadieron validaciones para la entidad vehiculos

+++ src/controllers/vehiculo.controller.ts
@@ -1,5 +1,5 @@
 import {VehiculoService} from "../services/vehiculo.service";
-import {PlacaDuplicada} from "../exceptions/vehiculo-excepciones";
+import {idNotFound, valorDuplicado} from "../exceptions/vehiculo-excepciones";
 
 let vehiculoService:VehiculoService;
 export class vehiculoController {
@@ -7,14 +7,12 @@ export class vehiculoController {
     constructor() {
         vehiculoService = new VehiculoService();
     }
-    async getVehiculos(req:any, res:any, next:any){
+    async getVehiculos(res:any, next:any){
         try {
-            const err = new PlacaDuplicada("La placa está duplicada", 400);
-            next(err);
             let vehiculos = await vehiculoService.getAll();
             res.status(200).json({status: 200, data: vehiculos, message: 'vehiculos encontrados'});
         }catch (error){
-            res.status(500).json({status: 500, message: error});
+            next(error);
         }
     }
 
@@ -22,20 +20,31 @@ export class vehiculoController {
         try {
             let id = parseInt(req.params.id);
             let vehiculo = await vehiculoService.getOne(id);
+            if(vehiculo === null){
+                throw new idNotFound('El id no existe')
+            }
             res.status(200).json({status: 200, data: vehiculo, message: 'vehiculo encontrados'});
         }catch (error){
-            res.status(500).json({status: 500, message: error});
+            next(error)
         }
     }
 
     async createVehiculo(req:any, res:any, next:any){
         try {
             let fecha_compra = new Date().toISOString();
+            let placaExiste = await vehiculoService.findPlaca(req.body.placa);
+            if(placaExiste){
+                throw new valorDuplicado('La placa esta duplicada');
+            }
+            let vinExiste = await vehiculoService.findVIN(req.body.VIN);
+            if(vinExiste){
+                throw new valorDuplicado('El VIN esta duplicado');
+            }
             let vehiculo = await vehiculoService.createVehiculo(req.body.placa, req.body.marca,
                 req.body.modelo, req.body.VIN, fecha_compra, req.body.costo, req.body.url_foto);
             res.status(200).json({status: 200, data: vehiculo, message: 'vehiculo creado'});
         }catch (error){
-            res.status(500).json({status: 500, message: error});
+            next(error);
         }
     }
 
@@ -43,11 +52,22 @@ export class vehiculoController {
         try {
             let id = parseInt(req.params.id);
             let vehiculo = await vehiculoService.getOne(id);
+            if(vehiculo === null){
+                throw new idNotFound('El id no existe')
+            }
+            let placaExiste = await vehiculoService.findPlaca(req.body.placa);
+            if(placaExiste){
+                throw new valorDuplicado('La placa esta duplicada');
+            }
+            let vinExiste = await vehiculoService.findVIN(req.body.VIN);
+            if(vinExiste){
+                throw new valorDuplicado('El VIN esta duplicado');
+            }
             vehiculo = await vehiculoService.updateVehiculos(id, req.body.placa, req.body.marca,
                 req.body.modelo, req.body.VIN, req.body.costo, req.body.url_foto);
             res.status(200).json({status: 200, data: vehiculo, message: 'vehiculo actualizado'});
         }catch (error){
-            res.status(500).json({status: 500, message: error});
+            next(error);
         }
     }
 
@@ -55,17 +75,20 @@ export class vehiculoController {
         try {
             let id = parseInt(req.params.id);
             let vehiculo = await vehiculoService.getOne(id);
+            if(vehiculo === null){
+                throw new idNotFound('El id no existe')
+            }
             vehiculo = await vehiculoService.deleteVehiculo(id)
             res.status(200).json({status: 200, data: vehiculo, message: 'vehiculo eliminado'});
         }catch (error){
-            res.status(500).json({status: 500, message: error});
+            next(error);
         }
     }
-    buildRequest(status:number, data:any, message:string){
+    /*buildRequest(status:number, data:any, message:string){
         return {
             status: status,
             'message': message,
             'data': data,
         };
-    }
+    }*/
 }
